+++ src/lib.rs
@@ -89,8 +89,6 @@ pub fn main() {
         EVENT_QUEUE.push(Event::smash_open());
     }
 
-    training::ui_hacks::install_hooks();
-
     hitbox_visualizer::hitbox_visualization();
     hazard_manager::hazard_manager();
     training::training_mods();
